backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
python-multipart
numpy
pandas
python-dotenv
pytest
httpx
requests 